refactor(webui): simplify form components
