+++ webui/src/app/(atlas)/[entry]/index.tsx
@@ -2,9 +2,9 @@ import { keepPreviousData, useQuery } from '@tanstack/react-query';
 
 import type { ModuleGraphResponse } from '~/app/--/entries/[entry]/modules/graph+api';
 import { BundleGraph } from '~/components/BundleGraph';
+import { BundleSelectForm } from '~/components/BundleSelectForm';
+import { ModuleFiltersForm } from '~/components/ModuleFilterForm';
 import { StateInfo } from '~/components/StateInfo';
-import { EntrySelectForm } from '~/components/forms/EntrySelect';
-import { ModuleFiltersForm } from '~/components/forms/ModuleFilter';
 import { EntryDeltaToast, useEntry } from '~/providers/entries';
 import { Layout, LayoutHeader, LayoutNavigation, LayoutTitle } from '~/ui/Layout';
 import { Spinner } from '~/ui/Spinner';
@@ -22,7 +22,7 @@ export default function BundlePage() {
   return (
     <Layout variant="viewport">
       <LayoutNavigation>
-        <EntrySelectForm />
+        <BundleSelectForm />
       </LayoutNavigation>
       <LayoutHeader>
         <LayoutTitle>
